:recycle: [#445] Properly convert the submission poll info and modal decorators to TS
